refactor(appui): extract transport-neutral client driver
